refactor(kamn-e2e-harness): consolidate docs-contract test files (#5690)
